openpgp: Align PacketPileParser::recursion_depth with PacketParser's

+++ openpgp/src/parse/packet_pile_parser.rs
@@ -218,9 +218,9 @@ impl<'a> PacketPileParser<'a> {
     ///
     /// A top-level packet has a recursion depth of 0.  Packets in a
     /// top-level container have a recursion depth of 1.  Etc.
-    pub fn recursion_depth(&self) -> Option<u8> {
+    pub fn recursion_depth(&self) -> Option<isize> {
         if let PacketParserResult::Some(ref pp) = self.ppr {
-            Some(pp.recursion_depth() as u8)
+            Some(pp.recursion_depth())
         } else {
             None
         }
